backend/song_recommendation_backend/requirements.txt
Django==4.0
djangorestframework==3.14
music-recommender
python-decouple==3.5
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.6.5
numpy==1.21.4
pandas==1.2.5
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
spotipy==2.19.0
textblob==0.17.1
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
seaborn
plotly
django-cors-headers